+++ src/app/admin/admin.module.ts
@@ -41,8 +41,8 @@ import { AdvertiserListItemComponent } from 'admin/users/advertiser-list/adverti
 import { PublisherListComponent } from 'admin/users/publisher-list/publisher-list.component';
 import { PublisherListItemComponent } from 'admin/users/publisher-list/publisher-list-item/publisher-list-item.component';
 import { PanelPlaceholdersSettingsComponent } from 'admin/panel-placeholders-settings/panel-placeholders-settings.component';
-import { BannerClassificationSettingComponent } from './general-settings/banner-classification-setting/banner-classification-setting/banner-classification-setting.component';
-import { MatRadioModule } from '@angular/material/radio'
+import { MatRadioModule } from '@angular/material/radio';
+import { SiteOptionsSettingsComponent } from './general-settings/site-options-settings/site-options-settings.component'
 
 @NgModule({
   imports: [
@@ -92,7 +92,7 @@ import { MatRadioModule } from '@angular/material/radio'
     RejectedDomainsComponent,
     PanelBlockadeComponent,
     PanelPlaceholdersSettingsComponent,
-    BannerClassificationSettingComponent,
+    SiteOptionsSettingsComponent,
   ],
   entryComponents: [
     SuccessSnackbarComponent,