+++ src/app/admin/general-settings/site-options-settings/site-options-settings.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core'
+import { Component, OnInit } from '@angular/core';
 import { AdminSiteOptions } from 'models/settings.model'
 import { Store } from '@ngrx/store'
 import { AppState } from 'models/app-state.model'
@@ -6,11 +6,11 @@ import { HandleSubscription } from 'common/handle-subscription'
 import {SetAdminSiteOptions } from 'store/admin/admin.actions'
 
 @Component({
-  selector: 'app-banner-classification-setting',
-  templateUrl: './banner-classification-setting.component.html',
-  styleUrls: ['./banner-classification-setting.component.scss']
+  selector: 'app-site-options-settings',
+  templateUrl: './site-options-settings.component.html',
+  styleUrls: ['./site-options-settings.component.scss']
 })
-export class BannerClassificationSettingComponent extends HandleSubscription implements OnInit {
+export class SiteOptionsSettingsComponent extends HandleSubscription implements OnInit {
   siteOptions: AdminSiteOptions
   canSubmit: boolean = false