+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Configuration default site classification and filtering  options on admin panel.
+- Configuration default site classification and filtering options on admin panel
 ### Changed
 - Wallet login buttons visibility
 - Angular 7